feat: translating the synopsis

+++ src/main/java/br/com/alura/screenMatch/model/SerieData.java
@@ -8,7 +8,7 @@ public record SerieData(
         @JsonAlias({"name", "title", "Title"}) String title,
         @JsonAlias({"synopsis", "Plot"}) String synopsis,
         @JsonAlias({"release_date", "first_air_date", "Released"}) String releaseDate,
-        @JsonAlias({"vote_average", "imdbRating"}) Float rating,
+        @JsonAlias({"vote_average", "imdbRating"}) Double rating,
         int totalSeasons,
         @JsonAlias("Genre") String genre,
         @JsonAlias("Actors") String actors,
